feat: add the real layouts' objectives to the 45 evaluation maps (#170)
* feat: add the real layouts' objectives to the 45 evaluation maps

feat: add the real layouts' objectives to the 45 evaluation maps

* feat: regenerate the map previews and add just render-maps

The 45 preview PNGs were produced ad hoc and committed without the code that
made them, so adding objectives left every one of them stale with nothing to
regenerate from. scripts/render_maps.py draws each preview from the map file
rather than from a reset episode: fifty deployed models cover the layout the
picture exists to show, and a seeded render would vary run to run.

Also re-measures the saturation note in configs/README.md against current main,
which moved four commits (real geometry, unit shooting) since it was written.


* feat: the objective is the ground — map objectives become their ruins

Every one of the 270 markers on the real layouts sits inside exactly one
terrain piece, so an objective is that piece's outline rather than a 3in disc
floating over the same footprint. This is the rules' terrain objective: you
hold it by standing on the ground.

The layouts print six markers, but on 24 of the 45 the two centre markers share
the board's largest ruin, and one piece of ground is held once — so those maps
carry five objectives and the other 21 carry six.

TerrainMapConfig's validator now requires each objective to be *determined*, an
area or x/y, rather than positioned.


* docs: an objective may be an area, not only x/y


---------

+++ wargame_rl/wargame/envs/types/config/entities.py
@@ -42,12 +42,12 @@ class ModelConfig(BaseModel):
 
     model_config = ConfigDict(extra="forbid")
 
-    x: int | None = Field(
+    x: float | None = Field(
         default=None,
         ge=0,
         description="X coordinate on the board. If None, placed randomly.",
     )
-    y: int | None = Field(
+    y: float | None = Field(
         default=None,
         ge=0,
         description="Y coordinate on the board. If None, placed randomly.",
@@ -93,12 +93,12 @@ class ObjectiveConfig(BaseModel):
 
     model_config = ConfigDict(extra="forbid")
 
-    x: int | None = Field(
+    x: float | None = Field(
         default=None,
         ge=0,
         description="X coordinate on the board. If None, placed randomly.",
     )
-    y: int | None = Field(
+    y: float | None = Field(
         default=None,
         ge=0,
         description="Y coordinate on the board. If None, placed randomly.",
